Refactor file distribution for Webauthn classes
